refactor: rename module names to avoid repetions in the file names
For example rename user-email file to email, because of this file placing in the user/module directory. It's obviosly, that this is a USER-email. But class name still USEREmail, because of we can have many emails. One contains in user/module, one more in the auth/module. And to avoid same class name still use different class names. (It's cool with vscode autocomplete import, we don't need for reading file name that imports, instead of class name.) And we can use UserEmail and AuthEmail in the one file, that one more reason to have different class names, but we still have same file names.